Add FIPS mode handling for pkcs11, tang-pin-TLS, and hybrid-TLS tests
Tests using RSA-PKCS (PKCS#1 v1.5) mechanism or ML-DSA-65
(post-quantum) algorithms fail in FIPS mode because these are
not FIPS-compliant.

- Add is_fips_enabled() helper to TestHelpers/utils.sh
- pkcs11/basic: skip test in FIPS mode (RSA-PKCS not allowed,
  SoftHSM lacks RSA-PKCS-OAEP support)
- tang-pin-TLS: skip pqc_alg variant in FIPS mode (ML-DSA-65
  not available)
- tang-pin-hybrid-TLS: gracefully degrade to 2-way hybrid
  (ECDSA+RSA) when ML-DSA-65 is unavailable in FIPS mode

+++ Sanity/tang-pin-TLS/runtest.sh
@@ -49,17 +49,28 @@ untrust_cert() {
 
 PACKAGE="clevis"
 
+SKIP_TEST=false
+
 rlJournalStart
     rlPhaseStartSetup
         rlRun ". ../../TestHelpers/utils.sh" || rlDie "cannot import function script"
         rlLog "TLS Certificate Algorithm: $CRYPTO_ALG"
         rlLog "TLS is ENABLED for this run."
-        rlRun "TmpDir=\$(mktemp -d)" 0 "Creating tmp directory"
-        rlRun "pushd $TmpDir"
-        rlRun "gen_tls_cert ${CRYPTO_ALG} 'server.key' 'server.crt'"
-        trust_cert
+
+        if is_fips_enabled && [ "$CRYPTO_ALG" = "ML-DSA-65" ]; then
+            rlLog "FIPS mode detected: ML-DSA-65 (post-quantum) is not available in FIPS mode, skipping test"
+            SKIP_TEST=true
+        fi
+
+        if ! $SKIP_TEST; then
+            rlRun "TmpDir=\$(mktemp -d)" 0 "Creating tmp directory"
+            rlRun "pushd $TmpDir"
+            rlRun "gen_tls_cert ${CRYPTO_ALG} 'server.key' 'server.crt'"
+            trust_cert
+        fi
     rlPhaseEnd
 
+  if ! $SKIP_TEST; then
     rlPhaseStart FAIL "tangd setup"
         # NOTE: This 'legacy_tang' check is from an old script version and will likely fail.
         # It's better to remove this if/else block if you only support modern tang.
@@ -103,11 +114,15 @@ CLEVIS_END
         rm -f adv plain enc plain2
     rlPhaseEnd
 
+  fi # ! $SKIP_TEST
+
     rlPhaseStartCleanup
-        stop_tang_fn "$port"
-        rlRun "popd"
-        untrust_cert
-        rlRun "rm -r $TmpDir" 0 "Removing tmp directory"
+        if ! $SKIP_TEST; then
+            stop_tang_fn "$port"
+            rlRun "popd"
+            untrust_cert
+            rlRun "rm -r $TmpDir" 0 "Removing tmp directory"
+        fi
     rlPhaseEnd
 rlJournalPrintText
 rlJournalEnd
